Remove stream size limit

+++ db/migrations/15_create_streams_global_index.sql
@@ -3,7 +3,7 @@ CREATE TABLE public.streams_global_index
     id                bigserial                         NOT NULL,
     partition_id      bigint                            NOT NULL,
     stream_id         character varying COLLATE "POSIX" NOT NULL,
-    stream_revision   integer                           NOT NULL,
+    stream_revision   bigint                            NOT NULL,
     starting_position bigint                            NOT NULL
 );
 
@@ -14,4 +14,4 @@ CREATE UNIQUE INDEX idx_streams_global_index_on_stream_id_and_partition_id ON pu
     USING btree (stream_id, partition_id);
 
 CREATE INDEX idx_streams_global_index_on_starting_position ON public.streams_global_index
-    USING btree(starting_position);
+    USING btree (starting_position);
